+++ jabgui/src/main/java/org/jabref/gui/preferences/websearch/WebSearchTabViewModel.java
@@ -142,6 +142,7 @@ public class WebSearchTabViewModel implements PreferenceTabViewModel {
         // add default search engines
         searchEngines.addAll(
                 new SearchEngineItem("Google Scholar", "https://scholar.google.com/scholar?q={title}"),
+                new SearchEngineItem("Semantic Scholar", "https://www.semanticscholar.org/search?q={title}"),
                 new SearchEngineItem("Short Science", "https://www.shortscience.org/internalsearch?q={title}")
         );
     }